加入 生成二维码

+++ common/src/main/java/com/pipi/common/persistence/mapper/FunMapper.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface FunMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(Fun record);
+    Long insert(Fun record);
 
     Fun selectByPrimaryKey(Long id);
